Finish renaming
